+++ R/shared-secret-filter.R
@@ -1,6 +1,6 @@
 #' @noRd
 sharedSecretFilter <- function(req, res){
-  secret <- getOption("plumber.sharedSecret", NULL)
+  secret <- get_option_or_env("plumber.sharedSecret", NULL)
   if (!is.null(secret)){
     supplied <- req$HTTP_PLUMBER_SHARED_SECRET
     if (!identical(supplied, secret)){
@@ -9,7 +9,7 @@ sharedSecretFilter <- function(req, res){
       res$serializer <- serializer_unboxed_json()
       # Using output similar to `defaultErrorHandler()`
       li <- list(error = "400 - Bad request")
-      
+
       # Don't overly leak data unless they opt-in
       if (is.function(req$pr$getDebug) && isTRUE(req$pr$getDebug())) {
         li$message <- "Shared secret mismatch"